rc1

+++ app/models/announcement.model.ts
@@ -1,13 +1,13 @@
 import { BelongsTo, Column, DataType, Model, Table } from "sequelize-typescript"
 import { User } from "@app/models/user.model"
-import { Field, Float, ObjectType } from "type-graphql"
+import { Field, Int, ObjectType } from "type-graphql"
 import { PartialUserBase } from "@app/classes/graphql/user/partialUser"
 import { DateType } from "@app/classes/graphql/serializers/date"
 
 @Table
 @ObjectType()
 export class Announcement extends Model {
-  @Field(() => Float)
+  @Field(() => Int)
   @Column({
     primaryKey: true,
     autoIncrement: true,
