+++ app/models/notification.model.ts
@@ -7,14 +7,14 @@ import {
   Table
 } from "sequelize-typescript"
 import { User } from "@app/models/user.model"
-import { Field, Float, ObjectType } from "type-graphql"
+import { Field, Int, ObjectType } from "type-graphql"
 import { PartialUserBase } from "@app/classes/graphql/user/partialUser"
 import { DateType } from "@app/classes/graphql/serializers/date"
 
 @ObjectType()
 @Table
 export class Notification extends Model {
-  @Field(() => Float)
+  @Field(() => Int)
   @Column({
     primaryKey: true,
     autoIncrement: true,
