+++ app/models/upload.model.ts
@@ -14,14 +14,14 @@ import { Collection } from "@app/models/collection.model"
 import { CollectionItem } from "@app/models/collectionItem.model"
 import { Star } from "@app/models/star.model"
 import { AutoCollectApproval } from "@app/models/autoCollectApproval.model"
-import { Field, Float, ObjectType } from "type-graphql"
+import { Field, Int, ObjectType } from "type-graphql"
 import { PartialUserBase } from "@app/classes/graphql/user/partialUser"
 import { DateType } from "@app/classes/graphql/serializers/date"
 
 @ObjectType()
 @Table
 export class Upload extends Model {
-  @Field(() => Float)
+  @Field(() => Int)
   @Column({
     primaryKey: true,
     autoIncrement: true,
